Renamed controller
