+++ config/routes.yaml
@@ -1,26 +1,26 @@
 pages_learn_classical:
     path:       /learn/classical
-    controller: App\Controller\Pages\Chessboard\ClassicalController::index
+    controller: App\Controller\Pages\Learn\ClassicalController::index
     methods:    GET
 
 pages_learn_fen_string:
     path:       /learn/fen-string
-    controller: App\Controller\Pages\Chessboard\FenStringController::index
+    controller: App\Controller\Pages\Learn\FenStringController::index
     methods:    GET
 
 pages_learn_fischer_random:
     path:       /learn/fischer-random
-    controller: App\Controller\Pages\Chessboard\FischerRandomController::index
+    controller: App\Controller\Pages\Learn\FischerRandomController::index
     methods:    GET
 
 pages_learn_rav_movetext:
     path:       /learn/rav-movetext
-    controller: App\Controller\Pages\Chessboard\RavMovetextController::index
+    controller: App\Controller\Pages\Learn\RavMovetextController::index
     methods:    GET
 
 pages_learn_san_movetext:
     path:       /learn/san-movetext
-    controller: App\Controller\Pages\Chessboard\SanMovetextController::index
+    controller: App\Controller\Pages\Learn\SanMovetextController::index
     methods:    GET
 
 pages_database_annotated_games:
