+++ src/Controller/Pages/Learn/FenStringController.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Controller\Pages\Chessboard;
+namespace App\Controller\Pages\Learn;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 
-class FischerRandomController extends AbstractController
+class FenStringController extends AbstractController
 {
     public function index(): Response
     {
-        return $this->render('pages/chessboard/fischer_random.html.twig');
+        return $this->render('pages/chessboard/fen_string.html.twig');
     }
 }
