+++ src/Controller/Pages/Learn/FischerRandomController.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Controller\Pages\Chessboard;
+namespace App\Controller\Pages\Learn;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 
-class SanMovetextController extends AbstractController
+class FischerRandomController extends AbstractController
 {
     public function index(): Response
     {
-        return $this->render('pages/chessboard/san_movetext.html.twig');
+        return $this->render('pages/chessboard/fischer_random.html.twig');
     }
 }
